+++ app/routes.php
@@ -18,7 +18,7 @@ $app->get('/{user}/list/{id}', 'ListController:index')->setName('list');
 // AUTH REQUIRED PAGES
 $app->group('', function () use ($app) {
 
-    $app->get('/auth/signout', 'AuthController:getSignOut')->setName('auth.signout');
+    $app->get('/auth/signout', 'SignOutController:getSignOut')->setName('auth.signout');
 
     // ACCOUNT ROUTES
     $app->get('/account/profile', 'ProfileSettingsController:getProfileSettings')->setName('account.profile');
@@ -54,11 +54,11 @@ $app->group('', function () use ($app) {
 $app->group('', function () use ($app) {
 
     // AUTH ROUTES
-    $app->get('/auth/signup', 'AuthController:getSignUp')->setName('auth.signup');
-    $app->post('/auth/signup', 'AuthController:postSignUp');
+    $app->get('/auth/signup', 'SignUpController:getSignUp')->setName('auth.signup');
+    $app->post('/auth/signup', 'SignUpController:postSignUp');
 
-    $app->get('/auth/signin', 'AuthController:getSignIn')->setName('auth.signin');
-    $app->post('/auth/signin', 'AuthController:postSignIn');
+    $app->get('/auth/signin', 'SignInController:getSignIn')->setName('auth.signin');
+    $app->post('/auth/signin', 'SignInController:postSignIn');
 
     $app->get('/auth/recover', 'RecoverController:getRecover')->setName('auth.recover');
     $app->post('/auth/recover', 'RecoverController:postRecover');
@@ -66,6 +66,6 @@ $app->group('', function () use ($app) {
     $app->post('/auth/reset/{identifier}', 'RecoverController:postReset');
 
 
-    $app->get('/auth/activate', 'AuthController:getActivate')->setName('auth.activate');
+    $app->get('/auth/activate', 'ActivateController:getActivate')->setName('auth.activate');
 
 })->add(new GuestMiddleware($container));